+++ metadata/tracking/letterboxd/example_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/golusoris/goenvoy/metadata/video/letterboxd"
+	"github.com/golusoris/goenvoy/metadata/tracking/letterboxd"
 )
 
 func Example() {